fix ProgramStub (bytes2hex were moved to the common project) #52
