+++ vm/src/test/scala/pravda/vm/ErrorTests.scala
