+++ vm/src/test/scala/pravda/vm/VmUtils.scala
