+++ vm/src/test/scala/pravda/vm/libs/TypedTests.scala
