+++ vm/src/test/scala/pravda/vm/opcode/BooleanOpcodesTests.scala
