+++ vm/src/test/scala/pravda/vm/opcode/ControlOpcodesTests.scala
