+++ vm/src/test/scala/pravda/vm/opcode/HeapOpcodesTests.scala
