+++ vm/src/test/scala/pravda/vm/opcode/LibraryTests.scala
