+++ vm/src/test/scala/pravda/vm/opcode/MathOpcodesTests.scala
