+++ vm/src/test/scala/pravda/vm/opcode/StackOpcodesTests.scala
