+++ vm/src/test/scala/pravda/vm/serialize/PushWordTests.scala
